# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书爆款笔记生成 agent Skill",
    "description": "Enterprise-grade hot-note research, multi-model generation, account analytics and controlled publishing Agent Skill.",
    "version": "5.14.5"
  },
  "paths": {
...
    "/v1/content/generate": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Generate",
        "operationId": "generate_v1_content_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryPackage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/content/generate-with-research": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Generate With Research",
        "operationId": "generate_with_research_v1_content_generate_with_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_with_research_v1_content_generate_with_research_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeliveryPackage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/content/generate-from-hot": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Generate From Hot",
        "description": "热门→选题→一键生成（dry_run 可只选题）。",
        "operationId": "generate_from_hot_v1_content_generate_from_hot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HotToNoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Generate From Hot V1 Content Generate From Hot Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/content/diagnose": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Diagnose",
        "operationId": "diagnose_v1_content_diagnose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/xhs_skill__api__routes__content__DiagnoseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Diagnose V1 Content Diagnose Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/content/rewrite": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Rewrite",
        "operationId": "rewrite_v1_content_rewrite_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewriteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Rewrite V1 Content Rewrite Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/content/reply-draft": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Reply Draft",
        "description": "生成授权评论回复草稿；auto_submit 恒为 false。",
        "operationId": "reply_draft_v1_content_reply_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Reply Draft V1 Content Reply Draft Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/v1/jobs/content/generate": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Submit Generate",
        "operationId": "submit_generate_v1_jobs_content_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_generate_with_research_v1_content_generate_with_research_post": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/GenerateRequest"
          },
          "report": {
            "$ref": "#/components/schemas/HotNotesReport"
          }
        },
        "type": "object",
        "required": [
          "request",
          "report"
        ],
        "title": "Body_generate_with_research_v1_content_generate_with_research_post"
      },
...
      "DeliveryPackage": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "assumptions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Assumptions"
          },
          "research_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Research Summary"
          },
          "hot_notes": {
            "items": {
              "$ref": "#/components/schemas/HotNoteCandidate"
            },
            "type": "array",
            "title": "Hot Notes"
          },
          "trend_insights": {
            "items": {
              "$ref": "#/components/schemas/TrendTopic"
            },
            "type": "array",
            "title": "Trend Insights"
          },
          "mechanisms": {
            "items": {
              "$ref": "#/components/schemas/ContentMechanism"
            },
            "type": "array",
            "title": "Mechanisms"
          },
          "strategy": {
            "additionalProperties": true,
            "type": "object",
            "title": "Strategy"
          },
          "title_candidates": {
            "items": {
              "$ref": "#/components/schemas/TitleCandidate"
            },
            "type": "array",
            "title": "Title Candidates"
          },
          "selected_title": {
            "type": "string",
            "title": "Selected Title"
          },
          "cover_options": {
            "items": {
              "$ref": "#/components/schemas/CoverOption"
            },
            "type": "array",
            "title": "Cover Options"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "graphic_pages": {
            "items": {
              "$ref": "#/components/schemas/GraphicPage"
            },
            "type": "array",
            "title": "Graphic Pages"
          },
          "video_script": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VideoScript"
              },
              {
                "type": "null"
              }
            ]
          },
          "media_assets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Media Assets"
          },
          "cover_asset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Asset"
          },
          "topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Topics"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "product_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Product Ids"
          },
          "keyword_map": {
            "additionalProperties": true,
            "type": "object",
            "title": "Keyword Map"
          },
          "hashtags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hashtags"
          },
          "pinned_comment": {
            "type": "string",
            "title": "Pinned Comment",
            "default": ""
          },
          "cta": {
            "type": "string",
            "title": "Cta",
            "default": ""
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/Claim"
            },
            "type": "array",
            "title": "Claims"
          },
          "originality_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Originality Report"
          },
          "compliance_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Compliance Report"
          },
          "ai_labeling": {
            "additionalProperties": true,
            "type": "object",
            "title": "Ai Labeling"
          },
          "quality_report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Quality Report"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "publication_status": {
            "type": "string",
            "title": "Publication Status",
            "default": "HUMAN_REVIEW_REQUIRED"
          },
          "human_review_required": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Human Review Required"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "trace_id",
          "selected_title",
          "body",
          "content_hash"
        ],
        "title": "DeliveryPackage"
      },
...
      "GenerateRequest": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Topic"
          },
          "objective": {
            "type": "string",
            "title": "Objective",
            "default": "search_growth"
          },
          "format": {
            "$ref": "#/components/schemas/ContentFormat",
            "default": "graphic"
          },
          "distribution_mode": {
            "$ref": "#/components/schemas/DistributionMode",
            "default": "hybrid"
          },
          "commercial_status": {
            "$ref": "#/components/schemas/CommercialStatus",
            "default": "NON_COMMERCIAL"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "product": {
            "additionalProperties": true,
            "type": "object",
            "title": "Product"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Evidence"
          },
          "brand_voice": {
            "additionalProperties": true,
            "type": "object",
            "title": "Brand Voice"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "research_current_trends": {
            "type": "boolean",
            "title": "Research Current Trends",
            "default": true
          },
          "web_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Web Results"
          },
          "candidate_count": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Candidate Count",
            "default": 6
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "suggested_topic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Topic"
          },
          "topic_angle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Angle"
          },
          "topic_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 400
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic Reason"
          },
          "note_style": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Style",
            "description": "review|seeding|avoid_pitfall|checklist|tutorial|store_visit|comparison|decision"
          },
          "narrative_framework": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Framework",
            "description": "pas|aida|bab|quest|four_p|scqa|auto"
          },
          "variant_index": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Variant Index",
            "default": 0
          },
          "video_duration_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60.0,
                "minimum": 15.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "topic"
        ],
        "title": "GenerateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotToNoteRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Query"
          },
          "suggestion_index": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Suggestion Index",
            "default": 0
          },
          "suggestion_topic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggestion Topic"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          },
          "use_account_health": {
            "type": "boolean",
            "title": "Use Account Health",
            "default": false
          },
          "format": {
            "type": "string",
            "pattern": "^(graphic|video)$",
            "title": "Format",
            "default": "graphic"
          },
          "video_duration_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 60.0,
                "minimum": 15.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Duration Seconds"
          },
          "note_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Style"
          },
          "narrative_framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Narrative Framework"
          },
          "target_audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Audience"
          },
          "commercial_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commercial Status"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "brand_voice": {
            "additionalProperties": true,
            "type": "object",
            "title": "Brand Voice"
          },
          "product": {
            "additionalProperties": true,
            "type": "object",
            "title": "Product"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Constraints"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "providers": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providers"
          },
          "web_results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Web Results"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "HotToNoteRequest",
        "description": "热门→选题→一键生成。"
      },
...
      "RewriteRequest": {
        "properties": {
          "body": {
            "type": "string",
            "title": "Body"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "commercial_status": {
            "type": "string",
            "title": "Commercial Status",
            "default": "NON_COMMERCIAL"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Constraints"
          },
          "references": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 200,
            "title": "References"
          }
        },
        "type": "object",
        "required": [
          "body"
        ],
        "title": "RewriteRequest"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}